+++ src/main/java/LadderGame/Main.java
@@ -1,5 +1,7 @@
 package LadderGame;
 
+import LadderGame.domain.Ladder;
+
 public class Main {
     public static void main(String[] args) {
         // 재귀 함수 시작 변수
